Config functions are now in a separate file. Added new quit phrase
